// src/app/auth/types/role-id.ts
/**
 * https://github.com/valitydev/bouncer-policies/blob/4bbeaef91653a40bf4583acc581bff1e5aba0ba6/policies/service/authz/roles/data.yaml#L15
 */
export enum RoleId {
    Administrator = 'Administrator',
    Manager = 'Manager',
    Accountant = 'Accountant',
    Integrator = 'Integrator',
    WalletManager = 'WalletManager',
}